Extract key mod
